% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/njord-package.R
\docType{package}
\name{njord-package}
\alias{njord}
\alias{njord-package}
\title{njord: Parametric Windstorm Trigger Design and Basis Risk Quantification}
\description{
Hazard modelling, parametric trigger design, and basis risk quantification
for UK/European windstorm using ERA5/Copernicus CDS data.
}
\author{
\strong{Maintainer}: Michael Campbell \email{m.campbell@tutanota.com}

Authors:
\itemize{
  \item Michael Campbell \email{m.campbell@tutanota.com}
}

}
\keyword{internal}
